Move base `ProviderApi` in its own file and rename providers' API files
