+++ src/h5web/providers/models.ts
@@ -1,3 +1,13 @@
+export interface ValueRequestParams {
+  path: string;
+  selection?: string;
+}
+
+export enum ProviderError {
+  NotFound = 'Entity not found',
+  Cancelled = 'Request cancelled',
+}
+
 /* -------------------- */
 /* ----- ENTITIES ----- */
 
@@ -158,8 +168,3 @@ export type Value<D extends Dataset> = D['shape'] extends ScalarShape
 
 export type H5WebComplex = [number, number];
 export type ComplexArray = (ComplexArray | H5WebComplex)[];
-
-export enum ProviderError {
-  NotFound = 'Entity not found',
-  Cancelled = 'Request cancelled',
-}
